Strategy pattern used
- Instead of having same file with 2 methods we used strategy pattern.
- Refactored into different files
