+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/images/preload/InAppImagePreloadConfig.kt
@@ -1,0 +1,13 @@
+package com.clevertap.android.sdk.inapp.images.preload
+
+internal data class InAppImagePreloadConfig(
+    val parallelDownloads: Int,
+) {
+    companion object {
+        private const val DEFAULT_PARALLEL_DOWNLOAD = 4
+
+        fun default() : InAppImagePreloadConfig = InAppImagePreloadConfig(
+                parallelDownloads = DEFAULT_PARALLEL_DOWNLOAD
+        )
+    }
+}
