+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/images/preload/InAppImagePreloaderCoroutine.kt
@@ -1,7 +1,7 @@
-package com.clevertap.android.sdk.inapp.images
+package com.clevertap.android.sdk.inapp.images.preload
 
 import com.clevertap.android.sdk.ILogger
-import com.clevertap.android.sdk.task.CTExecutors
+import com.clevertap.android.sdk.inapp.images.InAppResourceProvider
 import com.clevertap.android.sdk.utils.CtDefaultDispatchers
 import com.clevertap.android.sdk.utils.DispatcherProvider
 import kotlinx.coroutines.CoroutineExceptionHandler
@@ -12,17 +12,16 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.launch
 
-internal class InAppImagePreloader @JvmOverloads constructor(
-    private val inAppImageProvider: InAppResourceProvider,
-    private val executor: CTExecutors,
-    private val logger: ILogger? = null,
-    private val dispatchers: DispatcherProvider = CtDefaultDispatchers(),
-    private val config: InAppImagePreloadConfig = InAppImagePreloadConfig.default()
-) {
+internal class InAppImagePreloaderCoroutine @JvmOverloads constructor(
+        override val inAppImageProvider: InAppResourceProvider,
+        override val logger: ILogger? = null,
+        private val dispatchers: DispatcherProvider = CtDefaultDispatchers(),
+        override val config: InAppImagePreloadConfig = InAppImagePreloadConfig.default()
+) : InAppImagePreloaderStrategy {
 
     private var job: Job? = null
 
-    fun preloadImages(urls: List<String>) {
+    override fun preloadImages(urls: List<String>) {
 
         val handler = CoroutineExceptionHandler { _, throwable ->
             logger?.verbose("Cancelled image pre fetch \n ${throwable.stackTrace}")
@@ -50,32 +49,7 @@ internal class InAppImagePreloader @JvmOverloads constructor(
         }
     }
 
-    fun preloadImagesExec(urls: List<String>) {
-
-        for (url in urls) {
-            val task = executor.ioTaskNonUi<Void>()
-
-            task.execute("tag") {
-                inAppImageProvider.fetchInAppImage(url)
-                null
-            }
-        }
-    }
-
-    fun cleanup() {
+    override fun cleanup() {
         job?.cancel()
-        //executor?.shutdown
-    }
-}
-
-internal data class InAppImagePreloadConfig(
-    val parallelDownloads: Int,
-) {
-    companion object {
-        private const val DEFAULT_PARALLEL_DOWNLOAD = 4
-
-        fun default() : InAppImagePreloadConfig = InAppImagePreloadConfig(
-            parallelDownloads = DEFAULT_PARALLEL_DOWNLOAD
-        )
     }
 }
